[MODUSERSKC-71]. Change log level and move models
