+++ src/test/java/org/folio/uk/base/BaseIntegrationTest.java
@@ -38,11 +38,11 @@ import org.folio.test.extensions.impl.KafkaTestExecutionListener;
 import org.folio.test.extensions.impl.KeycloakExecutionListener;
 import org.folio.test.extensions.impl.WireMockAdminClient;
 import org.folio.test.extensions.impl.WireMockExecutionListener;
+import org.folio.uk.base.model.IdentityProvider;
 import org.folio.uk.domain.dto.User;
 import org.folio.uk.integration.keycloak.KeycloakClient;
 import org.folio.uk.integration.keycloak.KeycloakService;
 import org.folio.uk.integration.keycloak.TokenService;
-import org.folio.uk.integration.keycloak.model.IdentityProvider;
 import org.folio.uk.integration.keycloak.model.KeycloakUser;
 import org.folio.uk.it.CreateUserVerifyDto;
 import org.folio.uk.support.TestValues;
